Added view for ONLY starring projects for tracks
